requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==1.0.1
langchain-text-splitters==0.3.0
langchain-core==0.3.63
groq==0.34.1
faiss-cpu==1.7.4
pypdf==3.17.4
streamlit==1.29.0
python-dotenv==1.0.0
sentence-transformers==5.1.2
torch==2.9.1
transformers==4.57.1
numpy==1.26.4
